Remove classname from suite and it to both passed and failed tests.
As part of #1828 classname attribute was added to testsuite. This causes
Jenkins xunit plugin to fail to process reports generated by newman.

The following links proves that classname should be added only for testcase:

* https://docs.microsoft.com/en-us/azure/devops/pipelines/tasks/test/publish-test-results?view=azure-devops&tabs=yaml&viewFallbackFrom=vsts#result-formats-mapping

* https://github.com/windyroad/JUnit-Schema/blob/master/JUnit.xsd

* http://maven.apache.org/surefire/maven-surefire-plugin/xsd/surefire-test-report.xsd

At the same time classname was added only to failing tests which causes
incorrect calculation of age for the case by the same xunit plugin. In
order to fix this the classname attribute was added to both failed and
passed tests.

The classname now also generated from collection name based on the Java
identifiers rules (https://docs.oracle.com/javase/specs/jls/se8/html/jls-3.html#jls-3.8).
This is better than constant name cause allows to visually separate
tests from different collections if we run several of them as part of
a single Job in Jenkins.