NEWMAN-75: Do not store memory heavy objects in summary - instead give options
We are introducing a new flag `--skip-execution-response` on enabling which the summary response
will not be included in the report.

Also when `cli` is introduced as only reporter in the list of reporters we won't be saving execution responses.
This fix will allow developers using newman to generate the report even for large memory objects, in this case
only the summary response won't be part of the report.